+++ db/migrations/000051_billing_email.down.sql
@@ -1,0 +1,3 @@
+ALTER TABLE member_onboarding.application
+    DROP COLUMN IF EXISTS has_billing_email,
+    DROP COLUMN IF EXISTS billing_email;